+++ src/provider/SemanticTokensProvider/ModuleVarSemantic.ts
@@ -4,7 +4,9 @@ import type { TSemanticTokensLeaf } from './tools';
 export function ModuleVarSemantic(ModuleVar: TModuleVar): TSemanticTokensLeaf[] {
     const Tokens: TSemanticTokensLeaf[] = [];
     const { ModuleValMap } = ModuleVar;
-    for (const { defRangeList, refRangeList } of ModuleValMap.values()) {
+    for (const [k, { defRangeList, refRangeList }] of ModuleValMap) {
+        if (k === 'CLIPBOARD' || k === 'CLIPBOARDALL') continue;
+
         for (const { range } of [...defRangeList, ...refRangeList]) {
             Tokens.push({
                 range,